Login done, dashboard won't load anymore?
We are currently having issues displaying our table view. It must be something in our main dashboard view controller? When we try to log in (Which we have disabled for testing purposes), the dashboard crashes the app. We know it must be something with the dashboard because we had the login working properly before we tried adding firebase utilization to our events.